Maximum restriction limit supported

+++ Carespace.Bot/Save/Data.cs
@@ -12,7 +12,7 @@ public sealed class Data
     public Dictionary<int, EventData> Events { get; set; } = new();
 
     [UsedImplicitly]
-    public Dictionary<long, ushort> Strikes { get; set; } = new();
+    public Dictionary<long, byte> Strikes { get; set; } = new();
 
     [UsedImplicitly]
     public Dictionary<int, MessageData> Messages { get; set; } = new();
